requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via -r requirements.in, flask, pip-tools
entrypoints==0.3          # via flake8
flake8==3.7.9             # via -r requirements.in
flask==1.1.1              # via -r requirements.in
humanize==2.0.0           # via -r requirements.in
idna==2.9                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.11.1            # via -r requirements.in, flask
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
pip-tools==4.5.1          # via -r requirements.in
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pygal==2.4.0              # via -r requirements.in
python-dateutil==2.8.1    # via -r requirements.in
python-dotenv==0.12.0     # via -r requirements.in
requests==2.23.0          # via -r requirements.in
six==1.14.0               # via pip-tools, python-dateutil
urllib3==1.25.8           # via requests
werkzeug==1.0.0           # via flask
